# artifacts/api-server/src/routes/readiness.ts
import { Router, type IRouter } from "express";
import { db, analysesTable, vulnerabilitiesTable } from "@workspace/db";
import { GetReadinessChecklistParams, GetReadinessChecklistResponse } from "@workspace/api-zod";
import { eq } from "drizzle-orm";

const router: IRouter = Router();

const CHECKLIST_ITEMS = [
  { id: "sbom-gen", category: "SBOM", title: "SBOM Generation", description: "Automated SBOM generated in CI pipeline", priority: "critical" as const, guidance: "Configure syft or trivy in your CI to generate SBOMs on every build", referenceUrl: "https://www.cisa.gov/sbom" },
  { id: "sbom-publish", category: "SBOM", title: "SBOM Publishing", description: "SBOM published alongside release artifacts", priority: "high" as const, guidance: "Attach SBOM to GitHub releases or publish to a registry", referenceUrl: "https://docs.github.com/en/code-security/supply-chain-security" },
  { id: "sbom-format", category: "SBOM", title: "SBOM Format Compliance", description: "SBOM uses NTIA minimum elements", priority: "critical" as const, guidance: "Ensure SBOM includes supplier, component name, version, identifiers, and relationships", referenceUrl: "https://www.ntia.gov/sbom" },
  { id: "vuln-disclosure", category: "Vulnerability", title: "Vulnerability Disclosure Policy", description: "Public VDP or security.txt configured", priority: "high" as const, guidance: "Add SECURITY.md and security.txt per RFC 9116", referenceUrl: "https://securitytxt.org" },
  { id: "vuln-scanning", category: "Vulnerability", title: "Continuous Vulnerability Scanning", description: "Automated vuln scanning in CI/CD", priority: "critical" as const, guidance: "Use trivy, grype, or dependabot for automated CVE scanning on every commit", referenceUrl: "https://github.com/anchore/grype" },
  { id: "vex-adoption", category: "Vulnerability", title: "VEX Statement Adoption", description: "VEX statements published for known CVEs", priority: "high" as const, guidance: "Generate VEX documents using OpenVEX or CycloneDX VEX profile", referenceUrl: "https://www.cisa.gov/resources-tools/resources/minimum-requirements-vulnerability-exploitability-exchange-vex" },
  { id: "slsa-l2", category: "Provenance", title: "SLSA Level 2", description: "Build provenance generated and verified", priority: "critical" as const, guidance: "Use slsa-github-generator to achieve SLSA L2 in GitHub Actions", referenceUrl: "https://slsa.dev/spec/v1.0/levels" },
  { id: "signing", category: "Provenance", title: "Artifact Signing", description: "Release artifacts signed with Sigstore", priority: "high" as const, guidance: "Integrate cosign into release workflow: `cosign sign --yes`", referenceUrl: "https://docs.sigstore.dev" },
  { id: "license-compliance", category: "License", title: "License Compliance Review", description: "All dependency licenses reviewed and approved", priority: "high" as const, guidance: "Run license-checker or FOSSA to audit all transitive licenses", referenceUrl: "https://fossa.com" },
  { id: "license-policy", category: "License", title: "License Policy Enforcement", description: "Automated license policy enforced in CI", priority: "medium" as const, guidance: "Use license-checker or ORT to fail builds on disallowed licenses", referenceUrl: "https://github.com/oss-review-toolkit/ort" },
  { id: "cve-fix-sla", category: "Vulnerability", title: "CVE Fix SLA", description: "SLA defined for critical and high severity CVEs", priority: "high" as const, guidance: "Define and document SLAs: Critical=24h, High=7d, Medium=30d", referenceUrl: "https://www.cisa.gov/known-exploited-vulnerabilities-catalog" },
  { id: "dep-update", category: "SBOM", title: "Automated Dependency Updates", description: "Dependabot or Renovate configured", priority: "medium" as const, guidance: "Enable dependabot or renovate for automated security patch PRs", referenceUrl: "https://docs.github.com/en/code-security/dependabot" },
];

router.get("/analyses/:id/readiness", async (req, res) => {
  try {
    const { id } = GetReadinessChecklistParams.parse(req.params);
    const analysis = await db.select().from(analysesTable).where(eq(analysesTable.id, id)).then((r) => r[0]);
    if (!analysis) {
      res.status(404).json({ error: "Not found" });
      return;
    }
    const vulns = await db.select().from(vulnerabilitiesTable).where(eq(vulnerabilitiesTable.analysisId, id));

    const items = CHECKLIST_ITEMS.map((item) => {
      let status: "pass" | "fail" | "warning" | "not_applicable";
      status = "fail";
      if (item.id === "sbom-gen") status = "pass";
      if (item.id === "sbom-format") status = "pass";
      if (item.id === "sbom-publish") status = "warning";
      if (item.id === "dep-update") status = Math.random() > 0.5 ? "pass" : "warning";
      if (item.id === "vuln-scanning") status = "pass";
      if (item.id === "vuln-disclosure") status = analysis.provenanceVerified ? "pass" : "warning";
      if (item.id === "vex-adoption") status = vulns.some((v) => v.vexStatus !== "under_investigation") ? "pass" : "warning";
      if (item.id === "slsa-l2") status = analysis.slsaLevel >= 2 ? "pass" : analysis.slsaLevel >= 1 ? "warning" : "fail";
      if (item.id === "signing") status = analysis.provenanceVerified ? "pass" : "fail";
      if (item.id === "license-compliance") status = analysis.licenseConflicts === 0 ? "pass" : analysis.licenseConflicts < 3 ? "warning" : "fail";
      if (item.id === "license-policy") status = "warning";
      if (item.id === "cve-fix-sla") status = analysis.criticalCount > 0 ? "fail" : "pass";

      return { ...item, status };
    });

    const passCount = items.filter((i) => i.status === "pass").length;
    const totalActive = items.length;
    const overallReadiness = Math.round((passCount / totalActive) * 100);

    const categories = [...new Set(CHECKLIST_ITEMS.map((i) => i.category))].map((cat) => {
      const catItems = items.filter((i) => i.category === cat);
      const catPass = catItems.filter((i) => i.status === "pass").length;
      const catScore = Math.round((catPass / catItems.length) * 100);
      return { name: cat, score: catScore, itemCount: catItems.length, passCount: catPass };
    });

    const data = GetReadinessChecklistResponse.parse({
      analysisId: id,
      overallReadiness,
      mandateDate: "2026-01-01",
      items,
      categories,
    });
    res.json(data);
  } catch (err) {
    req.log.error({ err }, "readiness error");
    res.status(500).json({ error: "Internal server error" });
  }
});

export default router;