# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chainmail – Unified open-source maintainer workbench API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analyses:
    get:
      operationId: listAnalyses
      tags: [analyses]
      summary: List recent repository analyses
      responses:
        "200":
          description: List of analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analysis"
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Trigger a new repository analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisBody"
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
  /analyses/{id}:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get a specific analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analysis detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisDetail"
...
  /analyses/{id}/licenses:
    get:
      operationId: getLicenseReport
      tags: [licenses]
      summary: License compliance report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: License report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseReport"
  /analyses/{id}/license-tree:
    get:
      operationId: getLicenseTree
      tags: [licenses]
      summary: Dependency license inheritance tree
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: License tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LicenseTreeNode"
...
  /analyses/{id}/provenance:
    get:
      operationId: getProvenance
      tags: [provenance]
      summary: Provenance and attestation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Provenance status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvenanceStatus"
  /analyses/{id}/readiness:
    get:
      operationId: getReadinessChecklist
      tags: [readiness]
      summary: 2026 SBOM mandate readiness checklist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Readiness checklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessReport"
...
components:
  schemas:
...
    Analysis:
      type: object
      properties:
        id:
          type: string
        repo:
          type: string
        version:
          type: string
        commitSha:
          type: string
        status:
          type: string
          enum: [pending, running, complete, failed]
        riskScore:
          type: number
        criticalCount:
          type: integer
        highCount:
          type: integer
        componentCount:
          type: integer
        licenseConflicts:
          type: integer
        scannedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - repo
        - status
        - riskScore
        - criticalCount
        - highCount
        - componentCount
        - licenseConflicts
        - scannedAt
        - createdAt
    AnalysisDetail:
      allOf:
        - $ref: "#/components/schemas/Analysis"
        - type: object
          properties:
            slsaLevel:
              type: integer
            provenanceVerified:
              type: boolean
            sbomFormat:
              type: string
            sbomVersion:
              type: string
    CreateAnalysisBody:
      type: object
      properties:
        repoUrl:
          type: string
        branch:
          type: string
      required: [repoUrl]
...
    LicenseReport:
      type: object
      properties:
        analysisId:
          type: string
        distribution:
          type: array
          items:
            type: object
            properties:
              license:
                type: string
              count:
                type: integer
              percentage:
                type: number
              category:
                type: string
                enum: [permissive, weak_copyleft, strong_copyleft, unknown]
            required: [license, count, percentage, category]
        conflicts:
          type: array
          items:
            $ref: "#/components/schemas/LicenseConflict"
        compatibility:
          type: string
          enum: [clean, warnings, conflicts]
      required: [analysisId, distribution, conflicts, compatibility]
    LicenseConflict:
      type: object
      properties:
        id:
          type: string
        packageName:
          type: string
        packageVersion:
          type: string
        license:
          type: string
        conflictType:
          type: string
        severity:
          type: string
          enum: [error, warning, info]
        description:
          type: string
      required: [id, packageName, license, conflictType, severity, description]
    LicenseTreeNode:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        license:
          type: string
        licenseCategory:
          type: string
          enum: [permissive, weak_copyleft, strong_copyleft, unknown]
        hasConflict:
          type: boolean
        children:
          type: array
          items:
            $ref: "#/components/schemas/LicenseTreeNode"
      required: [name, version, license, licenseCategory, hasConflict]
...
    ProvenanceStatus:
      type: object
      properties:
        analysisId:
          type: string
        signed:
          type: boolean
        signingTool:
          type: string
        attestations:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              verified:
                type: boolean
              detail:
                type: string
            required: [type, verified, detail]
        slsaLevel:
          type: integer
        steps:
          type: array
          items:
            type: object
            properties:
              step:
                type: string
              status:
                type: string
                enum: [complete, incomplete, not_started]
              guidance:
                type: string
            required: [step, status, guidance]
      required:
        - analysisId
        - signed
        - attestations
        - slsaLevel
        - steps
    ReadinessReport:
      type: object
      properties:
        analysisId:
          type: string
        overallReadiness:
          type: number
        mandateDate:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReadinessItem"
        categories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              score:
                type: number
              itemCount:
                type: integer
              passCount:
                type: integer
            required: [name, score, itemCount, passCount]
      required:
        - analysisId
        - overallReadiness
        - mandateDate
        - items
        - categories
    ReadinessItem:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pass, fail, warning, not_applicable]
        priority:
          type: string
          enum: [critical, high, medium, low]
        guidance:
          type: string
        referenceUrl:
          type: string
      required:
        - id
        - category
        - title
        - description
        - status
        - priority
        - guidance
...